backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
PyJWT==2.9.0
reportlab==4.2.2
XlsxWriter==3.2.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.32

